+++ nfc/apdu.go
@@ -200,17 +200,18 @@ func DESFireWrapAPDU(cmd byte, data []byte) []byte {
 
 // DESFire native command codes
 const (
-	DFCmdSelectApplication = 0x5A
-	DFCmdGetApplicationIDs = 0x6A
-	DFCmdGetFileIDs        = 0x6F
-	DFCmdReadData          = 0xBD
-	DFCmdWriteData         = 0x3D
-	DFCmdAuthenticate      = 0x0A // Legacy DES auth
-	DFCmdAuthenticateISO   = 0x1A // 3DES auth
-	DFCmdAuthenticateAES   = 0xAA // AES auth
-	DFCmdGetVersion        = 0x60
-	DFCmdGetFileSettings   = 0xF5
-	DFCmdAdditionalFrame   = 0xAF
+	DFCmdSelectApplication  = 0x5A
+	DFCmdGetApplicationIDs  = 0x6A
+	DFCmdGetFileIDs         = 0x6F
+	DFCmdReadData           = 0xBD
+	DFCmdWriteData          = 0x3D
+	DFCmdAuthenticate       = 0x0A // Legacy DES auth
+	DFCmdAuthenticateISO    = 0x1A // 3DES auth
+	DFCmdAuthenticateAES    = 0xAA // AES auth
+	DFCmdGetVersion         = 0x60
+	DFCmdGetFileSettings    = 0xF5
+	DFCmdChangeFileSettings = 0x5F
+	DFCmdAdditionalFrame    = 0xAF
 )
 
 // DESFireSelectAppAPDU returns APDU for selecting a DESFire application